implement Watchdog Timer to avoid system hang, solving issue of not able to wake up properly form sleep mode
sporadically the board could not wake up from sleep mode. DDuring debug I discovered, that in fact the avr does wake up, but gets stuck later in the code, when the MPU is initialized. If the MPU fails to init, it will enter into a endless loop, causing a system hang.
The new implementation enables the watchdog reset with a timer of 8sec. In every loop the watchdog is patted (i.e. resetted). Prior o going into sleep mode, the watchdog is disabled not to wake up the saber automatically after 8sec. Then upon wake up the watchdog is enabled again. Even if the MPU hangs, the watchdog will issue a system reset if not patted, bringing out the system from a deadlock.

Sleep mode timer is set to 5mins, a more application friendly time.